refactor: Refactor CompanyDataStorage imports and improve handling
Replaced individual CompanyDataStorage imports with grouped imports from utils/storage to streamline and simplify code. Added a new method, hasSiret, and improved data validity checks to ensure consistent and accurate data handling throughout the application. These changes improve maintainability and functionality in storage management.
