+++ apps/web/src/WebApp.vue
@@ -33,7 +33,7 @@ import { useNavigationStore } from './stores/navigation'
 import TeeFooter from './components/TeeFooter.vue'
 import Translation from './utils/translation'
 import Cookie from './utils/cookies'
-import CompanyDataStorage from './utils/storage/companyDataStorage'
+import { CompanyDataStorage } from './utils/storage'
 
 const navigationStore = useNavigationStore()
 const router = useRouter()
