+++ apps/web/src/components/questionnaire/result/list/ResultHeader.vue
@@ -24,7 +24,7 @@
 import TrackStructure from '@/utils/track/trackStructure'
 import Translation from '@/utils/translation'
 import UsedTrack from '@/utils/track/usedTrack'
-import CompanyDataStorage from '@/utils/storage/companyDataStorage'
+import { CompanyDataStorage } from '@/utils/storage'
 
 const resume: string = Translation.t('programResults.resume', {
   effectif: Translation.t('enterprise.structureSize.' + (TrackStructure.getSize() ?? CompanyDataStorage.getSize() ?? '')),
