+++ apps/web/src/components/questionnaire/result/list/ResultList.vue
@@ -12,7 +12,7 @@ import { Project as ProjectType } from '@/types'
 import { Project } from '@/utils/project/project'
 import { useProjectStore } from '@/stores/project'
 import { Theme } from '@/utils/theme'
-import CompanyDataStorage from '@/utils/storage/companyDataStorage'
+import { CompanyDataStorage } from '@/utils/storage'
 
 const projectStore = useProjectStore()
 const navigationStore = useNavigationStore()
