+++ apps/web/src/components/questionnaire/result/list/ResultListInTabs.vue
@@ -42,7 +42,7 @@ import { storeToRefs } from 'pinia'
 import { computed } from 'vue'
 import { useProjectStore } from '@/stores/project'
 import { Theme } from '@/utils/theme'
-import CompanyDataStorage from '@/utils/storage/companyDataStorage'
+import { CompanyDataStorage } from '@/utils/storage'
 
 const navigationStore = useNavigationStore()
 const programStore = useProgramStore()
