+++ apps/web/src/components/questionnaire/result/list/ResultProjectList.vue
@@ -88,7 +88,7 @@ import OtherProjectCta from '@/components/project/list/OtherProjectCta.vue'
 import OtherProjectForm from '@/components/project/list/OtherProjectForm.vue'
 import Translation from '@/utils/translation'
 import Navigation from '@/utils/navigation'
-import CompanyDataStorage from '@/utils/storage/companyDataStorage'
+import { CompanyDataStorage } from '@/utils/storage'
 import { useNavigationStore } from '@/stores/navigation'
 
 interface ProjectListProps {
