+++ apps/web/src/components/track/TrackContent.vue
@@ -107,8 +107,7 @@ import TrackColOption from '@/utils/track/TrackColOption'
 import TrackComponent from '@/utils/track/TrackComponent'
 import { computed, ref } from 'vue'
 import { useRouter } from 'vue-router'
-import CompanyDataStorage from '@/utils/storage/companyDataStorage'
-import { CompanyDataStorageHandler } from '@/utils/storage/companyDataStorageHandler'
+import { CompanyDataStorage, CompanyDataStorageHandler } from '@/utils/storage'
 
 interface Props {
   trackElement: Element
@@ -219,7 +218,7 @@ const backToPreviousTrack = async () => {
 }
 
 watch(registeredData.value, async () => {
-  if (!CompanyDataStorage.hasData().value) return
+  if (!CompanyDataStorage.hasSiret()) return
 
   const next = CompanyDataStorageHandler.getNextTrackStorage()
