+++ apps/web/src/utils/storage/index.ts
@@ -1,0 +1,2 @@
+export * from './companyDataStorage'
+export * from './companyDataStorageHandler'
